+++ src/main/java/nedelosk/modularmachines/api/modular/module/tool/producer/energy/IProducerEngine.java
@@ -1,15 +1,48 @@
 package nedelosk.modularmachines.api.modular.module.tool.producer.energy;
 
+import nedelosk.modularmachines.api.modular.machines.basic.IModular;
+import nedelosk.modularmachines.api.modular.module.basic.IModule;
 import nedelosk.modularmachines.api.modular.module.tool.producer.IProducer;
+import nedelosk.modularmachines.api.modular.module.tool.producer.machine.IProducerMachine;
+import nedelosk.modularmachines.api.modular.utils.ModularUtils;
+import nedelosk.modularmachines.api.modular.utils.ModuleStack;
+import nedelosk.modularmachines.api.recipes.IRecipeManager;
+import nedelosk.modularmachines.api.recipes.RecipeInput;
 
 public interface IProducerEngine extends IProducer {
 
-	int getSpeedModifier(int tier);
+	int getSpeedModifier(ModuleStack stack);
+	
+	int getMaterialModifier(ModuleStack stack);
 
-	String getMode();
+	String getType();
+	
+	IRecipeManager getManager(ModuleStack stack);
+	
+	void setManager(IRecipeManager manager);
 
 	float getProgress();
 
 	void setProgress(float progress);
+	
+	int getBurnTime(ModuleStack stack);
+	
+	void setBurnTime(int burnTime);
+	
+	int getBurnTimeTotal(ModuleStack stack);
+	
+	void setBurnTimeTotal(int burnTimeTotal);
+	
+	boolean isWorking();
+	
+	void setIsWorking(boolean isWorking);
+	
+	IRecipeManager creatRecipeManager(IModular modular, String recipeName, int materialModifier, RecipeInput[] inputs);
+	
+	IRecipeManager creatRecipeManager();
+	
+	int getBurnTimeTotal(IModular modular, int speedModifier, ModuleStack<IModule, IProducerMachine> stackMachine, ModuleStack<IModule, IProducerEngine> stackEngine);
+	
+	int getBurnTimeTotal(IModular modular, ModuleStack<IModule, IProducerMachine> stackMachine, ModuleStack<IModule, IProducerEngine> stackEngine);
 
 }