+++ Changelog.md
@@ -1,6 +1,16 @@
 # Changelog
 
-# 0.3.2c
+# 0.3.3 Beta - Nedelosk
+
+**Modular-Machines**
+- Improve IProducerMachine, IProducerMachineRecipe, IProducerEngine API
+- Improve Engine and Recipe System
+- set Producer Boiler to machine
+
+*Plugins*
+- Improve Waila Plugin
+
+# 0.3.2c Beta - Nedelosk
 
 **Forest-Day**
 - Fix the Campfire Curb Item